+++ shui/api/install.py
