+++ shui/api/remote.py
@@ -2,7 +2,7 @@ import re
 import requests
 from bs4 import BeautifulSoup
 from url_normalize import url_normalize
-from shifl.version import Version
+from shui.version import Version
 
 
 def match_links(url, compiled_regex):
